Moved userdata structs out of cloudinit package so they can be reused by bottle rocket
